Putting pbc code in a separate module
