+++ qiskit/transpiler/preset_passmanagers/pbc/__init__.py
@@ -1,0 +1,29 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2026.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Transpilation into PBC (Pauli Based Computation)."""
+
+from .pbc_plugins import (
+    PBCUnrollPassManager,
+    PBCOptimizePassManager,
+    PBCTranslateToPBCPassManager,
+    PBCOptimizePBCPassManager,
+)
+from .pbc_pass_manager import generate_preset_pbc_pass_manager
+
+__all__ = [
+    "PBCOptimizePBCPassManager",
+    "PBCOptimizePassManager",
+    "PBCTranslateToPBCPassManager",
+    "PBCUnrollPassManager",
+    "generate_preset_pbc_pass_manager",
+]
